Abstract base class for live-reload ITs with global lock
